+++ engines/protocol/app/controllers/protocol/document_types_controller.rb
@@ -5,17 +5,17 @@ module Protocol
         before_action :set_document_type, only: [:edit, :destroy, :update]
 
         def index
-           # authorize @document_type
+           authorize @document_type
         end
 
         def new
             @document_type = DocumentType.new
-           # authorize @document_type
+            authorize @document_type
         end
 
         def create
             @document_type = DocumentType.new(document_type_params)
-            #authorize @document_type
+            authorize @document_type
             @document_type.save
         end
 
@@ -23,12 +23,12 @@ module Protocol
         end
 
         def update
-            #authorize @document_type
+            authorize @document_type
             @document_type.update(document_type_params)
         end
 
         def destroy
-           # authorize @document_type
+            authorize @document_type
             if @document_type.destroy
                 redirect_to action: 'index'
             end
